WIP/Discussion: Let server check an (optional) condition if it should continue to loop.
Instead of an external invocation of `stop()` to stop a running server, one could pass an lambda that returns true when the server should continue to run or false if it should stop.
Ideally this function should return as fast as possible and only do minimal amount of work. Like an interrupt function in µC.
This would ease the implementation of simple stopping mechanism that do not require external threads or other management.
For more complex functions, the minimal `bool()` function can be bound with `std::bind` to pass parameters etc.