Let the ConnectionHandlers count how many threads they spawn.
To reliable stop a ConnectionHandler it needs to count its spawned threads.
Because we can have multiple ConnectionHandlers, this counting can not be done via the `Components` since they *could* be shared between multiple ConnectionHandlers.
Thus we have to pass the counter separately.
While the counting can easily be done in the constructor and destructor of the Async-Api's `Coroutine` since it's threadpool cleanly allocs and deallocs the `Coroutine` the Simple-Api needs to be counted in the `Task::run()` method. It seems that std::thread does not deallocates the passed `HttpProcessor::Task` as expected (i.E. when the thread is done).