Restructured `ConsistentOutputStream` to not being related to `BufferedOutputStream`
`ConsistendOutputStream` is a promise. And this promise can only be kept by "not limited" buffers. In other Words, only buffers that can grow. When a consistend-sized buffer is full, it cant hold its `ConsistendOutputStream` promise. Both our output-streams - the `BufferOutputStream`, and the `ChunkedBuffer` - hold this promise. But it could be, that future `BufferedOutputStream` implementations have a fixed-size buffer and therefore do not hold this promise.