Refactored protocol-version evaluation to own `ConnectionHandler` (`HttpVersionSwitch`)
Handling switching to HTTP2 with an upgrade-handler interceptor was not feasible anymore.
Since we moved to async and an executor handles the rest, `HandleConnection` returns immediately.
Thus, the HTTP connection handler always answered with some generic response to the peer and broke the HTTP2 stream.
Now we have a designated `ConnectionHandler` for this which allows for more control.

The `Http2ConnectionHandler` is implemented in a way which allows it to be chained behind the `HttpVersionSwitch` or directly be called from an `ConnectionProvider`. It carries its own PRI-Message evaluation in an async context.